fix bug #5 8 threads per site did not work in timeslice pattern. Resolved by changing short mu, short id to integers. No idea why shorts break the code. Stupid behaviour can be reproduced like this: in LandauCoulombGaugeType::gatherInfo print the mu variable inside and outside of the if: if mu != 0 it will still evaluate to mu==0...
There is still one bug: On my GTX 580 sometimes I get: ERROR:
TimesliceGaugeFixingOverrelaxation(DoMicro) parity = 1: an illegal
memory access was encountered for the following settings in CoulombGauge
for the MicroKernel: GFLB(64,4)/8/0 
I did not see this problem on GTX 680 and GTX Titan

+++ include/gaugefixing/GaugeFixing4Threads.h
@@ -126,8 +126,8 @@ public:
 	}
 private:
 	static const int threadsPerSite = 4;
-	const short id;
-	const short mu;
+	const int id;
+	const int mu;
 	Algorithm algorithm;
 	LocalLinkType localLinkUp;
 	LocalLinkType localLinkDown;
